# docs/swagger.yaml
definitions:
  echo.HTTPError:
    properties:
      message: {}
    type: object
...
  github_com_schedule-rsreu_schedule-api_internal_models.AuditoriumSchedule:
    properties:
      auditorium:
        $ref: '#/definitions/github_com_schedule-rsreu_schedule-api_internal_models.Auditorium'
      denominator_period:
        example: 09.06-15.06
        type: string
      input_week_type:
        example: numerator
        type: string
      numerator_period:
        example: 16.06-22.06
        type: string
      schedule:
        $ref: '#/definitions/github_com_schedule-rsreu_schedule-api_internal_models.NumeratorDenominator-github_com_schedule-rsreu_schedule-api_internal_models_AuditoriumWeek'
    type: object
...
  github_com_schedule-rsreu_schedule-api_internal_models.CourseFaculties:
    properties:
      course:
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        example: 1
        type: integer
      faculties:
        items:
          enum:
          - иэф
          - фаиту
          - фвт
          - фрт
          - фэ
          type: string
        type: array
    type: object
  github_com_schedule-rsreu_schedule-api_internal_models.CourseFacultyGroups:
    properties:
      course:
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        example: 1
        type: integer
      faculty:
        enum:
        - иэф
        - фаиту
        - фвт
        - фрт
        - фэ
        example: фвт
        type: string
      groups:
        items:
          type: string
        type: array
    type: object
  github_com_schedule-rsreu_schedule-api_internal_models.Day:
    properties:
      day:
        enum:
        - Monday
        - Tuesday
        - Wednesday
        - Thursday
        - Friday
        - Saturday
        - Sunday
        example: Monday
        type: string
      day_ru:
        enum:
        - Пн
        - Вт
        - Ср
        - Чт
        - Пт
        - Сб
        - Вс
        example: Пн
        type: string
      time:
        example: "08.10"
        type: string
      week_type:
        enum:
        - числитель
        - знаменатель
        example: знаменатель
        type: string
      week_type_eng:
        enum:
        - numerator
        - denominator
        example: numerator
        type: string
    type: object
...
  github_com_schedule-rsreu_schedule-api_internal_models.Faculties:
    properties:
      faculties:
        items:
          type: string
        type: array
    type: object
...
  github_com_schedule-rsreu_schedule-api_internal_models.FacultyCourses:
    properties:
      courses:
        items:
          type: integer
        type: array
      faculty:
        enum:
        - иэф
        - фаиту
        - фвт
        - фрт
        - фэ
        example: фвт
        type: string
    type: object
  github_com_schedule-rsreu_schedule-api_internal_models.LessonType:
    properties:
      description:
        example: лабораторная
        type: string
      type:
        example: lab
        type: string
    type: object
...
  github_com_schedule-rsreu_schedule-api_internal_models.StudentSchedule:
    properties:
      course:
        example: 1
        type: integer
      denominator_period:
        example: 09.06-15.06
        type: string
      faculty:
        example: фвт
        type: string
      group:
        example: "344"
        type: string
      input_week_type:
        example: numerator
        type: string
      lessons_times:
        items:
          type: string
        type: array
      numerator_period:
        example: 16.06-22.06
        type: string
      schedule:
        $ref: '#/definitions/github_com_schedule-rsreu_schedule-api_internal_models.NumeratorDenominator-github_com_schedule-rsreu_schedule-api_internal_models_StudentWeek'
    type: object
...
  github_com_schedule-rsreu_schedule-api_internal_models.TeacherSchedule:
    properties:
      denominator_period:
        example: 09.06-15.06
        type: string
      departments:
        items:
          $ref: '#/definitions/github_com_schedule-rsreu_schedule-api_internal_models.Department'
        type: array
      full_name:
        example: Конюхов Алексей Николаевич
        type: string
      id:
        example: 1
        type: integer
      input_week_type:
        example: numerator
        type: string
      lessons_times:
        items:
          type: string
        type: array
      link:
        example: https://rsreu.ru/faculties/faitu/kafedri/vm/prepodavateli/9402-item-9402
        type: string
      numerator_period:
        example: 16.06-22.06
        type: string
      schedule:
        $ref: '#/definitions/github_com_schedule-rsreu_schedule-api_internal_models.NumeratorDenominator-github_com_schedule-rsreu_schedule-api_internal_models_TeacherWeek'
      short_name:
        example: Конюхов А.Н.
        type: string
    type: object
...
  internal_http_handlers_v1.schedulesByGroupsRequest:
    properties:
      groups:
        example:
        - "344"
        - "345"
        - "346"
        items:
          type: string
        type: array
    required:
    - groups
    type: object
...
info:
  contact: {}
  description: API for RSREU schedule.
  title: Schedule API
  version: "2.0"
paths:
  /api/v1/schedule/auditoriums:
    get:
      description: Get auditorium schedule by auditorium_id
      parameters:
      - description: auditorium_id
        example: 12
        in: query
        name: auditorium_id
        required: true
        type: integer
      - description: date
        example: "2025-06-13"
        in: query
        name: date
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_schedule-rsreu_schedule-api_internal_models.AuditoriumSchedule'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get auditorium schedule
      tags:
      - Auditoriums
  /api/v1/schedule/courses:
    get:
      description: Курсы факультета
      parameters:
      - description: faculty
        enum:
        - иэф
        - фаиту
        - фвт
        - фрт
        - фэ
        in: query
        name: faculty
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_schedule-rsreu_schedule-api_internal_models.FacultyCourses'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get faculty courses
      tags:
      - Courses
  /api/v1/schedule/day:
    get:
      description: Информация о текущем дне
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_schedule-rsreu_schedule-api_internal_models.Day'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get day
      tags:
      - Day
  /api/v1/schedule/faculties:
    get:
      description: Факультеты
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_schedule-rsreu_schedule-api_internal_models.Faculties'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get faculties
      tags:
      - Faculties
  /api/v1/schedule/faculties/course:
    get:
      description: Факультеты курса
      parameters:
      - description: course
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        in: query
        name: course
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_schedule-rsreu_schedule-api_internal_models.CourseFaculties'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get course faculties
      tags:
      - Faculties
  /api/v1/schedule/faculties/courses:
    get:
      description: Факультеты с курсами
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_schedule-rsreu_schedule-api_internal_models.FacultyCourses'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get faculties with courses
      tags:
      - Faculties
  /api/v1/schedule/groups:
    get:
      description: Группы факультета курса
      parameters:
      - description: course
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        in: query
        name: course
        type: integer
      - description: faculty
        enum:
        - иэф
        - фаиту
        - фвт
        - фрт
        - фэ
        in: query
        name: faculty
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_schedule-rsreu_schedule-api_internal_models.CourseFacultyGroups'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get course faculty groups
      tags:
      - Groups
  /api/v1/schedule/groups/{group}:
    get:
      description: Get schedule by group
      parameters:
      - description: group
        example: "344"
        in: path
        name: group
        required: true
        type: string
      - description: add empty lessons
        in: query
        name: add_empty_lessons
        type: boolean
      - description: date
        example: "2025-07-13"
        in: query
        name: date
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_schedule-rsreu_schedule-api_internal_models.StudentSchedule'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get schedule by group
      tags:
      - Groups
  /api/v1/schedule/groups/sample:
    post:
      description: Рассписание для нескольких групп
      parameters:
      - description: groups
        in: body
        name: groups
        required: true
        schema:
          $ref: '#/definitions/internal_http_handlers_v1.schedulesByGroupsRequest'
      - description: date
        example: "2025-07-13"
        in: query
        name: date
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_schedule-rsreu_schedule-api_internal_models.StudentSchedule'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get schedules by groups
      tags:
      - Groups
  /api/v1/schedule/lesson/types:
    get:
      description: Get lesson types
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_schedule-rsreu_schedule-api_internal_models.LessonType'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get lesson types
      tags:
      - Lesson
  /api/v1/schedule/teachers:
    get:
      description: Расписание преподавателя
      parameters:
      - description: teacher
        in: query
        name: teacher_id
        required: true
        type: integer
      - description: date
        example: "2025-07-13"
        in: query
        name: date
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_schedule-rsreu_schedule-api_internal_models.TeacherSchedule'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get teacher schedule
      tags:
      - Teachers
  /api/v1/schedule/teachers/all:
    get:
      description: Список всех преподавателей
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_schedule-rsreu_schedule-api_internal_models.TeachersList'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get teachers
      tags:
      - Teachers
  /api/v1/schedule/teachers/departments:
    get:
      description: Список кафедр. Если факультет не передана, то возвращаются все
        кафедры.
      parameters:
      - description: faculty_id
        example: 1
        in: query
        name: faculty_id
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_schedule-rsreu_schedule-api_internal_models.Department'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get departments list by faculty
      tags:
      - Teachers
  /api/v1/schedule/teachers/faculties:
    get:
      description: Список факультетов. Если кафедра не передан, то возвращаются все
        факультеты.
      parameters:
      - description: department_id
        example: 123
        in: query
        name: department_id
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_schedule-rsreu_schedule-api_internal_models.Faculty'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get faculties list by department
      tags:
      - Teachers
  /api/v1/schedule/teachers/list:
    get:
      description: Список преподавателей по факультету и кафедре. Параметры не обязательны.
      parameters:
      - description: faculty_id
        example: 4
        in: query
        name: faculty_id
        type: integer
      - description: department_id
        example: 17
        in: query
        name: department_id
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_schedule-rsreu_schedule-api_internal_models.TeachersList'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get teachers list by faculty and department
      tags:
      - Teachers
swagger: "2.0"